Make RequestLogFactory a SPI
Move the SLF4J implentation to a separate class.

Closes #1290
