+++ dropwizard-core/src/main/java/io/dropwizard/server/AbstractServerFactory.java
@@ -25,6 +25,7 @@ import io.dropwizard.jetty.GzipFilterFactory;
 import io.dropwizard.jetty.MutableServletContextHandler;
 import io.dropwizard.jetty.NonblockingServletHolder;
 import io.dropwizard.jetty.RequestLogFactory;
+import io.dropwizard.jetty.Slf4jRequestLogFactory;
 import io.dropwizard.lifecycle.setup.LifecycleEnvironment;
 import io.dropwizard.servlets.ThreadNameFilter;
 import io.dropwizard.util.Duration;
@@ -203,7 +204,7 @@ public abstract class AbstractServerFactory implements ServerFactory {
 
     @Valid
     @NotNull
-    private RequestLogFactory requestLog = new RequestLogFactory();
+    private RequestLogFactory requestLog = new Slf4jRequestLogFactory();
 
     @Valid
     @NotNull
