+++ dropwizard-jetty/src/test/java/io/dropwizard/jetty/RequestLogFactoryTest.java
@@ -17,7 +17,7 @@ import java.util.TimeZone;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class RequestLogFactoryTest {
-    private RequestLogFactory requestLog;
+    private Slf4jRequestLogFactory slf4jRequestLog;
 
     @Before
     public void setUp() throws Exception {
@@ -25,7 +25,7 @@ public class RequestLogFactoryTest {
         objectMapper.getSubtypeResolver().registerSubtypes(ConsoleAppenderFactory.class,
                                                            FileAppenderFactory.class,
                                                            SyslogAppenderFactory.class);
-        this.requestLog = new ConfigurationFactory<>(RequestLogFactory.class,
+        this.slf4jRequestLog = new ConfigurationFactory<>(Slf4jRequestLogFactory.class,
                                                      BaseValidator.newValidator(),
                                                      objectMapper, "dw")
                 .build(new File(Resources.getResource("yaml/requestLog.yml").toURI()));
@@ -33,7 +33,7 @@ public class RequestLogFactoryTest {
 
     @Test
     public void defaultTimeZoneIsUTC() {
-        assertThat(requestLog.getTimeZone())
+        assertThat(slf4jRequestLog.getTimeZone())
             .isEqualTo(TimeZone.getTimeZone("UTC"));
     }
 }
